rename package to match repo
also rename applicationId in build.gradle
