+++ SourceEditor/app/src/androidTest/java/com/microsoft/device/display/wm_samples/sourceeditor/SourceEditorInDualScreenTest.kt
@@ -1,4 +1,4 @@
-package com.microsoft.device.display.samples.sourceeditor
+package com.microsoft.device.display.wm_samples.sourceeditor
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
@@ -7,11 +7,11 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.filters.MediumTest
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner
 import androidx.test.rule.ActivityTestRule
-import com.microsoft.device.display.samples.sourceeditor.utils.childOf
-import com.microsoft.device.display.samples.sourceeditor.utils.setOrientationLeft
-import com.microsoft.device.display.samples.sourceeditor.utils.setOrientationRight
-import com.microsoft.device.display.samples.sourceeditor.utils.switchFromSingleToDualScreen
-import com.microsoft.device.display.samples.sourceeditor.utils.unfreezeRotation
+import com.microsoft.device.display.wm_samples.sourceeditor.utils.childOf
+import com.microsoft.device.display.wm_samples.sourceeditor.utils.setOrientationLeft
+import com.microsoft.device.display.wm_samples.sourceeditor.utils.setOrientationRight
+import com.microsoft.device.display.wm_samples.sourceeditor.utils.switchFromSingleToDualScreen
+import com.microsoft.device.display.wm_samples.sourceeditor.utils.unfreezeRotation
 import org.hamcrest.Matchers
 import org.junit.After
 import org.junit.Rule
