+++ SourceEditor/app/src/androidTest/java/com/microsoft/device/display/wm_samples/sourceeditor/utils/SurfaceDuoUtils.kt
@@ -4,7 +4,7 @@
  *
  */
 
-package com.microsoft.device.display.samples.sourceeditor.utils
+package com.microsoft.device.display.wm_samples.sourceeditor.utils
 
 import android.view.Surface
 import androidx.test.platform.app.InstrumentationRegistry
