+++ SourceEditor/app/src/androidTest/java/com/microsoft/device/display/wm_samples/sourceeditor/utils/TestUtils.kt
@@ -1,4 +1,4 @@
-package com.microsoft.device.display.samples.sourceeditor.utils
+package com.microsoft.device.display.wm_samples.sourceeditor.utils
 
 import android.view.View
 import androidx.test.espresso.ViewAction
