+++ SourceEditor/app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.microsoft.device.display.samples.sourceeditor"
+    package="com.microsoft.device.display.wm_samples.sourceeditor"
     tools:ignore="GoogleAppIndexingWarning">
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -22,7 +22,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
-        <activity android:name="com.microsoft.device.display.samples.sourceeditor.MainActivity"
+        <activity android:name="com.microsoft.device.display.wm_samples.sourceeditor.MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
